Added Timer.h
Simple timer that is easy to use and uses an example of variadic templates in C++. Code is portable for measuring any function passed into it via function pointer and its arguments in a comma separated list. There is (still) old code in here, clean up is needed (mostly old ID for the timers).