Added print, insert, and substring functions
Still working on printf() and seeing if there is a better way to go about it. printf() will be using the insert() method.
For conveniences sake, I switched over to Semantic versioning with a descriptor as to what build is presented. (dev = in development, release = official release, debug = debugging enabled)